dynamic quantization working
