+++ machop/chop/passes/graph/analysis/__init__.py
@@ -31,3 +31,7 @@ from .pruning import (
     add_natural_sparsity_metadata_analysis_pass,
     hook_inspection_analysis_pass,
 )
+
+from .runtime import (
+    runtime_analysis_pass
+)
