+++ machop/chop/passes/graph/analysis/runtime/__init__.py
@@ -1,1 +1,1 @@
-from .analysis import RuntimeAnalysis
+from .analysis import runtime_analysis_pass
